Test restructure step 6 PR-A: relocate 2 misfiled tests
PreviewAppServerStreamTests moves from IOSAppServerTests.swift
(MCPIntegrationTests, a real-daemon target) to a new file in
PreviewsIOSTests (the matching unit target): it's already fully
fake-based (StubFrameSource/NoopInputSink), no daemon or simulator
dependency, just misfiled in an integration-test target.

withTimeoutFiresUnderStarvation moves from MacOSMCPTests.swift to a new
MCPTestServerHarnessTests.swift in the same target: it tests the test
harness's MCPTestServer.withTimeout, not product code, so it doesn't
belong mixed into the real-daemon integration suite.

Zero coverage or logic change — both are straight relocations, verified
byte-for-byte identical test bodies.
